refactoring:the auth component was moved due to improve the current code
