+++ wolox-books/src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { RegisterComponent } from './screens/auth/screens/register/register.component';
-import { LoginComponent } from './screens/auth/screens/login/login.component';
+import { RegisterComponent } from './screens/register/register.component';
+import { LoginComponent } from './screens/login/login.component';
 import { AuthComponent } from './screens/auth/auth.component';
 
 const routes: Routes = [
