+++ wolox-books/src/app/app.module.ts
@@ -4,9 +4,9 @@ import { NgModule } from '@angular/core';
 import {FormsModule, ReactiveFormsModule} from '@angular/forms';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { RegisterComponent } from './screens/auth/screens/register/register.component';
+import { RegisterComponent } from './screens/register/register.component';
 import { HttpClientModule } from '@angular/common/http';
-import { LoginComponent } from './screens/auth/screens/login/login.component';
+import { LoginComponent } from './screens/login/login.component';
 import { AuthComponent } from './screens/auth/auth.component';  
 
 @NgModule({
