+++ wolox-books/src/app/screens/login/login.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
-import { UserService } from '../../../../services/user.service';
+import { UserService } from '../../services/user.service';
 import { Router } from '@angular/router';
-import { LocalStorageService } from '../../../../services/local-storage.service';
+import { LocalStorageService } from '../../services/local-storage.service';
 
 @Component({
   selector: 'app-login',
