+++ wolox-books/src/app/screens/register/register.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../variables.scss';
+@import '../../../variables.scss';
 
 .main-page {
     display: flex;
